requirements.txt
streamlit
openai
pandas
python-docx
PyMuPDF
tiktoken
